Massive move of functions to helpers
