+++ src/staging/staging_helpers.py
@@ -1,5 +1,18 @@
 import pandas as pd
 from numpy import random
+import logging
+import re
+from datetime import datetime
+from typing import Callable
+
+from src.utils.wrappers import time_logger_wrap
+from src.staging import validation as val
+from src.staging import spp_parser, history_loader
+from src.staging import spp_snapshot_processing as processing
+
+
+# Create logger for this module
+StagingHelperLogger = logging.getLogger(__name__)
 
 
 def postcode_topup(mystr: str, target_len: int = 8) -> str:
@@ -129,3 +142,299 @@ def update_ref_list(full_df: pd.DataFrame, ref_list_df: pd.DataFrame) -> pd.Data
     df = df.drop(["_merge", "cellnumber_new"], axis=1)
 
     return df
+
+
+def getmappername(mapper_path_key, split):
+
+    patt = re.compile(r"^(.*?)_path")
+    mapper_name = re.search(patt, mapper_path_key).group(1)
+
+    if split:
+        mapper_name = mapper_name.replace("_", " ")
+
+    return mapper_name
+
+
+def load_valdiate_mapper(
+    mapper_path_key,
+    paths,
+    file_exists_func,
+    read_csv_func,
+    logger,
+    val_with_schema_func: Callable,
+    one_to_many_val_func: Callable,
+    *args,
+):
+    """
+    Loads a specified mapper, validates it using a schema and an optional validation function.
+
+    This function first retrieves the path of the mapper from the provided paths dictionary using the mapper_path_key.
+    It then checks if the file exists at the mapper path. If the file exists, it is read into a DataFrame.
+    The DataFrame is then validated against a schema, which is located at a path constructed from the mapper name.
+    If a validation function is provided, it is called with the DataFrame and any additional arguments.
+
+    Args:
+        mapper_path_key (str): The key to retrieve the mapper path from the paths dictionary.
+        paths (dict): A dictionary containing paths.
+        file_exists_func (Callable): A function to check if a file exists at a given path.
+        read_csv_func (Callable): A function to read a CSV file into a DataFrame.
+        logger (logging.Logger): A logger to log information and errors.
+        val_with_schema_func (Callable): A function to validate a DataFrame against a schema.
+        validation_func (Callable, optional): An optional function to perform additional validation on the DataFrame.
+        *args: Additional arguments to pass to the validation function.
+
+    Returns:
+        pd.DataFrame: The loaded and validated mapper DataFrame.
+
+    Raises:
+        FileNotFoundError: If no file exists at the mapper path.
+        ValidationError: If the DataFrame fails schema validation or the validation function.
+    """
+    # Get the path of the mapper from the paths dictionary
+    mapper_path = paths[mapper_path_key]
+
+    # Get the name of the mapper from the mapper path key
+    mapper_name = getmappername(mapper_path_key, split=True)
+
+    # Log the loading of the mapper
+    logger.info(f"Loading {getmappername(mapper_path_key, split=True)} to File...")
+
+    # Check if the file exists at the mapper path, raise an error if it doesn't
+    file_exists_func(mapper_path, raise_error=True)
+
+    # Read the file at the mapper path into a DataFrame
+    mapper_df = read_csv_func(mapper_path)
+
+    # Construct the path of the schema from the mapper name
+    schema_prefix = "_".join(word for word in mapper_name.split() if word != "mapper")
+    schema_path = f"./config/{schema_prefix}_schema.toml"
+
+    # Validate the DataFrame against the schema
+    val_with_schema_func(mapper_df, schema_path)
+
+    # If a one-to-many validation function is provided, validate the DataFrame
+    if one_to_many_val_func:
+        # Prepend the DataFrame to the arguments
+        args = (mapper_df,) + args
+        # Call the validation function with the DataFrame and the other arguments
+        one_to_many_val_func(*args)  # args include "col_many" and "col_one"
+
+    # Log the successful loading of the mapper
+    logger.info(f"{mapper_name} loaded successfully")
+
+    # Return the loaded and validated DataFrame
+    return mapper_df
+
+
+def load_historic_data(config: dict, paths: dict, read_csv: Callable) -> dict:
+    """Load historic data into the pipeline.
+
+    Args:
+        config (dict): The pipeline configuration
+        paths (dict): The paths to the data files
+        read_csv (Callable): Function to read a csv file.
+            This will be the hdfs or network version depending on settings.
+
+    Returns:
+        dict: A dictionary of history data loaded into the pipeline.
+    """
+    curent_year = config["years"]["current_year"]
+    years_to_load = config["years"]["previous_years_to_load"]
+    years_gen = history_loader.history_years(curent_year, years_to_load)
+
+    if years_gen is None:
+        StagingHelperLogger.info("No historic data to load for this run.")
+        return {}
+    else:
+        StagingHelperLogger.info("Loading historic data...")
+        history_path = paths["history_path"]
+        dict_of_hist_dfs = history_loader.load_history(
+            years_gen, history_path, read_csv
+        )
+        # Check if it has loaded and is not empty
+        if isinstance(dict_of_hist_dfs, dict) and bool(dict_of_hist_dfs):
+            StagingHelperLogger.info(
+                "Dictionary of history data: %s loaded into pipeline",
+                ", ".join(dict_of_hist_dfs),
+            )
+            StagingHelperLogger.info("Historic data loaded.")
+        else:
+            StagingHelperLogger.warning(
+                "Problem loading historic data. Dict may be empty or not present"
+            )
+            raise Exception("The historic data did not load")
+
+    return dict_of_hist_dfs if dict_of_hist_dfs else {}
+
+
+def check_snapshot_feather_exists(
+    config: dict,
+    check_file_exists: Callable,
+    feather_file_to_check,
+    secondary_feather_file,
+) -> bool:
+    """Check if one or both of snapshot feather files exists.
+
+    Conifg arguments decide whether to check for one or both.
+
+    Args:
+        config (dict): The pipeline configuration
+        check_file_exists (Callable): Function to check if file exists
+            This will be the hdfs or network version depending on settings.
+
+    Returns:
+        bool: True if the feather file exists, False otherwise.
+    """
+
+    if config["global"]["load_updated_snapshot"]:
+        return check_file_exists(feather_file_to_check) and check_file_exists(
+            secondary_feather_file
+        )
+    else:
+        return check_file_exists(feather_file_to_check)
+
+
+@time_logger_wrap
+def load_snapshot_feather(feather_file, read_feather):
+    snapdata = read_feather(feather_file)
+    StagingHelperLogger.info(f"{feather_file} loaded")
+    return snapdata
+
+
+def load_val_snapshot_json(snapshot_path, load_json, config, network_or_hdfs):
+
+    StagingHelperLogger.info("Loading SPP snapshot data from json file")
+
+    # Load data from JSON file
+    snapdata = load_json(snapshot_path)
+
+    contributors_df, responses_df = spp_parser.parse_snap_data(snapdata)
+
+    # Get response rate
+    res_rate = "{:.2f}".format(processing.response_rate(contributors_df, responses_df))
+
+    # the anonymised snapshot data we use in the DevTest environment
+    # does not include the instance column. This fix should be removed
+    # when new anonymised data is given.
+    if network_or_hdfs == "hdfs" and config["global"]["dev_test"]:
+        responses_df = fix_anon_data(responses_df, config)
+    StagingHelperLogger.info("Finished Data Ingest...")
+
+    # Validate snapshot data
+    val.validate_data_with_schema(contributors_df, "./config/contributors_schema.toml")
+    val.validate_data_with_schema(responses_df, "./config/long_response.toml")
+
+    # Data Transmutation
+    full_responses = processing.full_responses(contributors_df, responses_df)
+
+    StagingHelperLogger.info(
+        "Finished Data Transmutation and validation of full responses dataframe"
+    )
+    # Validate and force data types for the full responses df
+    # TODO Find a fix for the datatype casting before uncommenting
+    val.combine_schemas_validate_full_df(
+        full_responses,
+        "./config/contributors_schema.toml",
+        "./config/wide_responses.toml",
+    )
+
+    return full_responses, res_rate
+
+
+def load_validate_secondary_snapshot(load_json, secondary_snapshot_path):
+
+    # Load secondary snapshot data
+    StagingHelperLogger.info("Loading secondary snapshot data from json file")
+    secondary_snapdata = load_json(secondary_snapshot_path)
+
+    # Parse secondary snapshot data
+    secondary_contributors_df, secondary_responses_df = spp_parser.parse_snap_data(
+        secondary_snapdata
+    )
+
+    # Validate secondary snapshot data
+    StagingHelperLogger.info("Validating secondary snapshot data...")
+    val.validate_data_with_schema(
+        secondary_contributors_df, "./config/contributors_schema.toml"
+    )
+    val.validate_data_with_schema(secondary_responses_df, "./config/long_response.toml")
+
+    # Create secondary full responses dataframe
+    secondary_full_responses = processing.full_responses(
+        secondary_contributors_df, secondary_responses_df
+    )
+    # Validate and force data types for the secondary full responses df
+    val.combine_schemas_validate_full_df(
+        secondary_full_responses,
+        "./config/contributors_schema.toml",
+        "./config/wide_responses.toml",
+    )
+
+    # return secondary_full_responses
+    return secondary_full_responses
+
+
+def write_snapshot_to_feather(
+    feather_dir_path: str,
+    snapshot_name: str,
+    full_responses: pd.DataFrame,
+    write_feather,
+    secondary_snapshot_name: str,
+    secondary_full_responses: pd.DataFrame = None,
+) -> None:
+    """
+    Writes the provided DataFrames to feather files.
+
+    This function writes the `full_responses` DataFrame to a feather file named
+    "{snapshot_name}_corrected.feather", and the `secondary_full_responses`
+    DataFrame to a feather file named "{secondary_snapshot_name}.feather".
+    Both files are written to the provided `feather_path`.
+
+    Args:
+        feather_path (str): The path where the feather files will be written.
+        snapshot_name (str): The name of the snapshot for the `full_responses`
+        DataFrame.
+        full_responses (pd.DataFrame): The DataFrame to write to the
+        "{snapshot_name}_corrected.feather" file.
+        secondary_snapshot_name (str): The name of the snapshot for the
+        `secondary_full_responses` DataFrame.
+        secondary_full_responses (pd.DataFrame): The DataFrame to write to the
+        "{secondary_snapshot_name}.feather" file.
+    """
+
+    feather_file = f"{snapshot_name}_corrected.feather"
+    feather_file_path = os.path.join(feather_dir_path, feather_file)
+    write_feather(feather_file_path, full_responses)
+    StagingHelperLogger.info(f"Written {feather_file} to {feather_dir_path}")
+
+    if secondary_full_responses is not None:
+        secondary_feather_file = os.path.join(
+            feather_dir_path, f"{secondary_snapshot_name}.feather"
+        )
+        write_feather(secondary_feather_file, secondary_full_responses)
+        StagingHelperLogger.info(
+            f"Written {secondary_snapshot_name}.feather to {feather_dir_path}"
+        )
+
+
+def stage_validate_harmonise_postcodes(
+    config, paths, full_responses, run_id, check_file_exists, read_csv, write_csv
+):
+    """
+    Stage, validate and harmonise the postcode column
+    """
+    StagingHelperLogger.info("Starting PostCode Validation")
+    postcode_masterlist = paths["postcode_masterlist"]
+    check_file_exists(postcode_masterlist, raise_error=True)
+    postcode_mapper = read_csv(postcode_masterlist)
+    postcode_masterlist = postcode_mapper["pcd2"]
+    invalid_df = val.validate_post_col(full_responses, postcode_masterlist, config)
+    StagingHelperLogger.info("Saving Invalid Postcodes to File")
+    pcodes_folder = paths["postcode_path"]
+    tdate = datetime.now().strftime("%Y-%m-%d")
+    invalid_filename = f"invalid_unrecognised_postcodes_{tdate}_v{run_id}.csv"
+    write_csv(f"{pcodes_folder}/{invalid_filename}", invalid_df)
+    StagingHelperLogger.info("Finished PostCode Validation")
+
+    return full_responses, postcode_mapper
+
